+++ src/zope/password/tests/test_zpasswd.py
@@ -50,9 +50,8 @@ class TestBase(unittest.TestCase):
 
     @contextlib.contextmanager
     def patched_stdio(self, input_data=None):
-        io_type = io.StringIO if bytes is not str else io.BytesIO
-        self.stdout = io_type()
-        self.stderr = io_type()
+        self.stdout = io.StringIO()
+        self.stderr = io.StringIO()
 
         self.old_stdout = sys.stdout
         self.old_stderr = sys.stderr
@@ -61,7 +60,7 @@ class TestBase(unittest.TestCase):
         sys.stderr = self.stderr
 
         if input_data is not None:
-            self.stdin = io_type(input_data)
+            self.stdin = io.StringIO(input_data)
             sys.stdin = self.stdin
 
         try:
@@ -126,7 +125,10 @@ class ArgumentParsingTestCase(TestBase):
         os.close(handle)
         self.addCleanup(os.remove, path)
         options = self.parse_args([option, path])
-        self.assertEqual(options.destination.name, path)
+        try:
+            self.assertEqual(options.destination.name, path)
+        finally:
+            options.destination.close()
 
     def test_destination_long(self):
         self.test_destination_short("--output")